// src/app/api/model/initiateFirstPensionBillResponseDTOJsonAPIResponse.ts
/**
 * CTS-BE
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { APIResponseStatus } from './aPIResponseStatus';
import { InitiateFirstPensionBillResponseDTO } from './initiateFirstPensionBillResponseDTO';


export interface InitiateFirstPensionBillResponseDTOJsonAPIResponse { 
    result?: InitiateFirstPensionBillResponseDTO;
    apiResponseStatus?: APIResponseStatus;
    message?: string | null;
}
export namespace InitiateFirstPensionBillResponseDTOJsonAPIResponse {
}