// src/app/api/model/ppoPaymentListItemDTO.ts
/**
 * CTS-BE
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DateOnly } from './dateOnly';


export interface PpoPaymentListItemDTO { 
    dataSource?: { [key: string]: any; } | null;
    fromDate?: DateOnly;
    toDate?: DateOnly;
    basicPensionAmount?: number;
    breakupId?: number;
    componentName?: string | null;
    componentType?: string;
    breakupAmount?: number;
    rateId?: number;
    rateType?: string;
    rateAmount?: number;
    periodInMonths?: number;
    dueAmount?: number;
}